refactor: move AddAccountUseCase to account folder
