+++ src/domain/usecases/authentication/authentication.ts
@@ -1,0 +1,3 @@
+export interface IAuthentication {
+  execute(email: string, password: string): Promise<string>
+}
