+++ src/domain/usecases/authentication/index.ts
@@ -1,2 +1,1 @@
-export * from './account'
 export * from './authentication'
