+++ src/presentation/controllers/sign-up/protocols.ts
@@ -1,3 +1,3 @@
 export * from '@/presentation/protocols'
 export * from '@/domain/models/account'
-export * from '@/domain/usecases/add-account'
+export * from '@/domain/usecases/account/add'
